% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data in shiny app}
\usage{
update_data(
  base_folder_path = NULL,
  spp_list_csv = NULL,
  spp_core_csv = NULL,
  project_status_rds = NULL,
  all_events_rds = NULL,
  all_counts_rds = NULL,
  locations_rds = NULL
)
}
\arguments{
\item{base_folder_path}{path to location of files}

\item{all_counts_rds}{}
}
\description{
Update data in shiny app
}
