% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.CWS.ONT-package.R
\docType{package}
\name{Shiny.CWS.ONT-package}
\alias{Shiny.CWS.ONT}
\alias{Shiny.CWS.ONT-package}
\title{\pkg{Shiny.CWS.ONT} R package (CWS Ontario Data graphical user interface)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Hope \email{david.hope@ec.gc.ca} (\href{https://orcid.org/0000-0002-2140-4261}{ORCID})

}
\keyword{internal}
