% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data in shiny app}
\usage{
update_data(
  spp_list_csv = NULL,
  spp_core_csv = NULL,
  project_status_rds = NULL,
  all_events_rds = NULL,
  all_counts_rds = NULL
)
}
\arguments{
\item{all_counts_rds}{}
}
\description{
Update data in shiny app
}
